requirements.txt
requests
websockets
websocket_client